+++ packages/editor/CHANGELOG.md
@@ -9,6 +9,7 @@
 - `Flex → Children Wrap` option change to be shown by default.
 - The `Purge Cache` button for style regeneration will remain active after purging the cache.
 - Improved cache validation and configuration updates for blockera extensions when options are modified.
+- Cleaned up cache data and improved cache functionality in block extensions for a better and easier user experience.
 
 ### Bug Fixes